创建新分支 直接在老项目基础上调试

+++ src/main/java/de/ostfalia/bips/ws24/camunda/database/domain/KategoriegewichtInProjekt.java
@@ -13,7 +13,7 @@ public class KategoriegewichtInProjekt {
     private Id id;
 
     @Column(name = "gewicht", nullable = false)
-    private float gewicht;
+    private Double gewicht;
 
     public KategoriegewichtInProjekt() {
         this.id = new Id();
@@ -23,7 +23,7 @@ public class KategoriegewichtInProjekt {
         this.id = id;
     }
 
-    public void setGewicht(float gewicht) {
+    public void setGewicht(double gewicht) {
         this.gewicht = gewicht;
     }
 
@@ -31,7 +31,7 @@ public class KategoriegewichtInProjekt {
         return id;
     }
 
-    public float getGewicht() {
+    public double getGewicht() {
         return gewicht;
     }
